Extracted a helper method ready for re-use.
