+++ src/helpers/editor.ts
@@ -1,6 +1,6 @@
 import i18n from "@/i18n";
 import { useStore } from "@/store/store";
-import {AddFrameCommandDef, AddShorthandFrameCommandDef, AllFrameTypesIdentifier, areSlotCoreInfosEqual, BaseSlot, CaretPosition, FieldSlot, FrameContextMenuActionName, FrameContextMenuShortcut, FramesDefinitions, getFrameDefType, isFieldBaseSlot, isFieldBracketedSlot, isFieldMediaSlot, isFieldStringSlot, isSlotBracketType, isSlotQuoteType, isSlotStringLiteralType, MediaSlot, ModifierKeyCode, NavigationPosition, Position, SelectAllFramesFuncDefScope, SlotCoreInfos, SlotCursorInfos, SlotsStructure, SlotType, StringSlot} from "@/types/types";
+import {AddFrameCommandDef, AddShorthandFrameCommandDef, AllFrameTypesIdentifier, areSlotCoreInfosEqual, BaseSlot, CaretPosition, CollapsedState, FieldSlot, FrameContextMenuActionName, FrameContextMenuShortcut, FrameObject, FramesDefinitions, getFrameDefType, isFieldBaseSlot, isFieldBracketedSlot, isFieldMediaSlot, isFieldStringSlot, isSlotBracketType, isSlotQuoteType, isSlotStringLiteralType, MediaSlot, ModifierKeyCode, NavigationPosition, Position, SelectAllFramesFuncDefScope, SlotCoreInfos, SlotCursorInfos, SlotsStructure, SlotType, StringSlot} from "@/types/types";
 import { getAboveFrameCaretPosition, getAllChildrenAndJointFramesIds, getAvailableNavigationPositions, getFrameBelowCaretPosition, getFrameContainer, getFrameSectionIdFromFrameId } from "./storeMethods";
 import { splitByRegexMatches, strypeFileExtension } from "./common";
 import {getContentForACPrefix} from "@/autocompletion/acManager";
@@ -2281,4 +2281,46 @@ export function slotStructureToString(ss: SlotsStructure) : string {
         r.push(getMatchingBracket(ss.openingBracketValue, true));
     }
     return r.join("");
+}
+
+// Given the current state of all the frames (or undefined if in a mixed state), gets the next state that would/should be cycled to
+// This is the next state that all frames support.
+export function calculateNextCollapseState(curCommonState : CollapsedState | undefined, frameList: FrameObject[]) : CollapsedState {
+    // The default if anything unexpected happens:
+    let nextState = CollapsedState.FULLY_VISIBLE;
+    // If they are in a mixed state we want next state to be fully collapsed, which is always available for defs:
+    if (curCommonState === undefined) {
+        nextState = CollapsedState.ONLY_HEADER_VISIBLE;
+    }
+    else {
+        // Otherwise, we need to work out the next state.  It should be a state they can all reach, which
+        // depends on the frames present (e.g. the doc state can't be reached if we have a mix of class and
+        // function frames)
+
+        // Step 1: compute remaining states per object
+        const possibleNextStates = (frameList as FrameObject[]).map((f) => {
+            const idx = f.frameType.allowedCollapsedStates.indexOf(f.collapsedState ?? CollapsedState.FULLY_VISIBLE);
+            if (idx < 0) {
+                return []; // safeguard if current not found
+            }
+
+            // everything after current + everything before current
+            return f.frameType.allowedCollapsedStates.slice(idx + 1).concat(f.frameType.allowedCollapsedStates.slice(0, idx));
+        });
+
+        // Step 2: intersect them all
+        let intersection = new Set(possibleNextStates[0]);
+        for (let i = 1; i < possibleNextStates.length; i++) {
+            intersection = new Set(possibleNextStates[i].filter((s) => intersection.has(s)));
+        }
+
+        // Step 3: pick earliest in canonical order (say, the first object’s list)
+        for (const state of possibleNextStates[0]) {
+            if (intersection.has(state)) {
+                nextState = state;
+                break;
+            }
+        }
+    }
+    return nextState;
 }
